btcxblockchainapi/requirements.txt
Django==2.0.2
djangorestframework
python-bitcoinrpc
unittest2
simplejson
ddt
pyyaml
web3

